docker/trend-qdrant/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2024.2.2
exceptiongroup==1.2.1
grpcio==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.7.0
numpy==1.20.3
pip==24.0
portalocker==2.7.0
protobuf==4.24.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
qdrant-client==1.4.0
setuptools==57.5.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.7.1
urllib3==1.26.18
wheel==0.37.1
zipp==3.15.0
